stdenv: Major rework
This is a fundamental change which aims to ensure our toolchain is fully
up to date and fully re-bootstraps itself. Our old toolchain actually
used the downloaded compiler and linker binaries to compile code used in
our final stdenv.

With this change we now have glibc 2.28. The only shortcoming of this
new glibc is that we don't have locale-archive patches yet.

Our stdenv has considerably fewer packages included in the build
routines.